dom/base/test/file_bug1263696_frame_fail.html

<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"
 "http://www.w3.org/TR/REC-html401-19991224/strict.dtd">
<html>
 <head>
  <title>Bug 1263696 - iframe that should not be loaded</title>
 </head>
 <body>
   <script>
    parent.SimpleTest.ok(false, "this iframe should not load");
   </script>
 </body>
</html>
